Added Heisenberg trotterization for a circuit (with error plot test) in trotter/ folder. Heisenberg model build hamiltonian (heisenberg/ folder). Important fix to gradient function (a conjugate was missing) so that local SVD update compression leads to monotonic loss. All test work, even for long circuits (this was an issue before the fix). Updated tests that were relying on old gradient function. Add 'normalize' method in MPO. Fix optimizer (pass left_to_right is now coherent with pass right_to_left). Add Weyl decomposition circuit from brickwall. (weyl_circuit_builder.py) Add archived.

+++ rqcopt_mpo/tensor_network/core_ops.py
@@ -278,7 +278,7 @@ def compress_SVD(
         return u, s, vh, 0
         
     # Determine truncation based on cutoff
-    k_cutoff = jnp.sum(s / s[0] > cutoff)
+    k_cutoff = int(jnp.sum(s / s[0] > cutoff))
 
     # Determine truncation based on max_bondim
     if max_bondim is None:
